pyproject.toml
[project]
name = "nellabot"
version = "0.1.0"
description = "Nella — always-on personal AI assistant with Telegram interface"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=22.0",
    "anthropic>=0.79.0",
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
    "mem0ai>=0.1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
